+++ io.openems.edge.evcs.webasto.unite/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>io.openems.edge.evcs.webasto</name>
+	<name>io.openems.edge.evcs.webasto.unite</name>
 	<comment></comment>
 	<projects>
 	</projects>
