+++ io.openems.edge.evcs.webasto.unite/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: OpenEMS Edge EVCS Webasto
+Bundle-Name: OpenEMS Edge EVCS Webasto Unite
 Bundle-Vendor: OpenEMS Hackathon
 Bundle-License: https://opensource.org/licenses/EPL-2.0
 Bundle-Version: 1.0.0.${tstamp}
